requirements.txt
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colored==1.4.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.27
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
